--- agents/MATLABagent_tcp/temp_simulation_script.m
@@ -1,14 +1,0 @@
-
-    params = struct(...
-        'timeStep', 0.1, ...
-        'maxTime', 10.0, ...
-        'initialPosition', [0, 0], ...
-        'initialVelocity', [1.0, 4.0], ...
-        'gravity', 9.81, ...
-        'mass', 1.0, ...
-        'logFrequency', 5, ...
-        'port', 12345, ...
-        'host', 'localhost' ...
-    );
-    simulation(params)
-
