--- src/simulation-bridge/config/config.yml
@@ -1,14 +1,0 @@
-mqtt:
-  broker: "test.mosquitto.org"
-  port: 1883
-simulations:
-  matlab:
-    input_topic: "sim/matlab/input"
-    output_topic: "sim/matlab/output"
-  analogic:
-    input_topic: "sim/analogic/input"
-    output_topic: "sim/analogic/output"
-  python:
-    input_topic: "sim/python/input"
-    output_topic: "sim/python/output"
-timeout: 10
